requirements.txt
streamlit
pymongo
pandas
Pillow
matplotlib